Move MQTT related classes to mqtt_utils for further modularization.
With this change, simulation_sm becomes an easier to read and edit file.
The intent is to have it be the main executable for the simulations.
As such, all classes were moved to their own script files.
Also, MQTT publisher and subscriber have been instantiated as a part
of the Scenario.