Update classes to the new data model and change "static" name to "stationary"
- This is a first attempt at implementing the new data model. With the new implementation,
some class methods (such as Robot.move()) needed some changes.
- StaticRobot was renamed to StationaryRobot, and its type was also changed to "stationary".
This new name just makes more sense.
- New class Header was created to avoid repeating code that every other class needs. This
contains basic information such as id, name, description, and some others.
- The MQTT topic is now automatically created by combining the base topic with the entity's
namespace and id. This is a first step into further modularizing the MQTT functionality, as
the idea is to not have any MQTT related attributes inside the Robot classes.
- Class inheritance now actually inherits from the parent class (that's what the super().__init__
is for).
- All Robot's mqtt_payload have been emptyed. It needs to be updated with the new datamodel,
but first I'd like to attempt implementing a new method for sending data to MQTT. This should
appear in a future commit. The AGV have "pose" in its mqtt_payload to show that the topics
are still working.
- Change status to uppercase so it's similar to how it's being handled inside ROS.
- Change Robot.move(): now it checks if the target has been reached before doing any
other checks and calculations. Also removed angle, since the new data model uses
orientation (a quaternion).