Update with new data structure and MQTT changes
- Add "_v2" classes for WIP classes with new data structure;
- Add new classes for Process Step and Operation;
- Remove methods and constant related to MQTT: the publishing functionality has
been completely modularized to mqtt_utils.py. Further discussion about these changes
can be found in the mqtt_util.py commit;
- Change the way dictionaries are copied: I found out that I was copying the dictionary ref
instead of creating a new copy. This caused changes to the new dict to be reflected on the
old one (for example, initial_pose was always updated with the new values for pose). To
fix this, I used the "deepcopy" method of the copy library, which does what we need.