Update to MQTT payload publishing and ShopfloorManager
- All MQTT publishing is now handled by objects inside mqtt_utils.py;
- ShopfloorPublisher now has a "send_payload" method, which
gets an entity's attributes and turns it into a JSON payload. This payload
is then sent to the entity's mqtt topic. With this, we are able to completely
modularize the MQTT functionality: entites.py classes no longer require
the mqtt_update_payload and mqtt_send_payload methods to work
with MQTT, as well as no longer requiring any mqtt related attributes.
Everything gets handled by the ShopfloorManager now;
- ShopfloorPublisher's send_payload() method includes edge cases
to avoid publishing repeated payloads;
- Update MqttGeneric init function to include the root_topic attribute;
- Removed the @staticmethod wrappers from the MQTT methods.
They weren't necessary and, with their removal, it's now possible to
call self;
- Implement next_action method in ShopfloorManager for returning
an action received via MQTT. These actions are now stored in a list
object that works like a queue and are used by the StateMachine
class in a scenario;
- Add is_jsonable method to check if an object is JSON serializable.