Add Job queue feature + Job manipulation via MQTT + new States
- Add a function to handle State transitions according to Job status,
and another to handle creating new Jobs and adding them to the queue
- Update the current State prints to the terminal: now if the same State
is repeated multiple times, instead of flooding the terminal, it'll be
printed in the same line with the number of times it repeated besides it
- Replace the action-based solution to Job manipulation with the Job
Manager approach
- Update to States to implement the new transitions according to Job
status: only the States in the "IN_PROGRESS" section of the Job Flow
(check the diagrams) received this update
- Update Idle State: it'll loop forever on itself unless a message with
"IN_PROGRESS" is sent to the Job's status topic.
- Add a OP00 State
- Add a FinishJob State: handles everything needed to be done before
creating a new Job
- Update Reset State: after resetting all entities in the resettable
entities lists, it'll create a new Job
- Update ProcessStep instantiation so it creates a deepcopy of the
needed Operation instead of getting its reference: this is needed when
creating new Jobs to avoid overwriting other Jobs PS or OPs
- Replace the Shopfloor.entities list with the "resettable_entities" and
"publishing_entities" list, since not all entities do both of resetting
and publishing. Also, Jobs, PS and OPs won't be reset anymore, since
Jobs will be created dynamically
- Update the subscriber's topic to include the '+' wildcard, allowing it
to watch all Job's statuses
- Delete old unused imports