Add MQTT Client for DTV use case: DTVMqttClient
- Add DTVMqttClient class to handle MQTT communication (sub/pub) in DTV
specific scenarios
- Add references to both current scenario and the scenario manager. This
allows checking and updating properties from both
- Add JobManager methods to DTVMqttClient. Note that it only checks the
name of the last subtopic in a subscribed topic. If 2 topics end the same,
this implementation will need to change (it was made like this because
the code looks cleaner like so)
- Change Structure list generation from for loops to list compreheension.
It's both faster and easier to read
- Change method name from "mqtt_loop" to "publish_thread"
- Change job_update_queue from being a MQTT property to being a scenario
property. This makes more sense, as methods related to Jobs will be handled
by the scenario. Now MQTT has a ref to the scenario so it can still
access the job_update_queue
- Move MQTT settings to settings.py