Add example and updated DTV scenarios
- Add new example scenarios. These show the barebones implementation
of the state machine
- Add ./dtv scenarios folder for DTV specific scenarios. flexibility0 and
flexibility1 are updated copies of the scenarioTwinViewer1 scenario,
adapted to fit the new scenario manager

- About the DTV scenario changes:
- Add reference to the scenario manager. This reference comes as a automatically
set parameter when initializing the scenario, but it needs to be assigned
to Scenario.manager
- Change print logs to use [#] for informational prints, [W] for warnings,
and [!] for errors
- Add SCENARIO_NAME constant to get the scenario name from the filename
- Add FLEXIBILITY constant to define the scenario's flexibility id
- Add Initialize and Shutdown states. MQTT is now defined inside the
Initialie state and shut down in the Shutdown state. Other initialization
stuff should go in this state. In the Shutdown state, sets the Scenario.is_active
flag to False to allow the State Machine to shutdown, returning to the
scenario manager
- Remove add_job and check_job_status_then_change_state functions. They're
now defined as methods in state_machine.SimulatedScenario
- Add check for changes in selected_flexibility in the check_job_status_then_change_state
method to allow transitions to the shutdown state
- Move "create 3 jobs" from global scope to the new Initialize State
- Add conditions to check for selected_flexibility change in the Idle
and OnHold states